Clonar repo ao usar o update.sh quando a pasta .git não for encontrada. Remover tentativa falha de usar quebra de linha nos arquivos sh...

+++ update.sh
@@ -2,8 +2,15 @@
 
 trap 'kill $(jobs -pr)' SIGINT SIGTERM EXIT
 
-git reset --hard
-git pull --allow-unrelated-histories -X theirs
+if [ ! -d ".git" ]; then
+  git init
+  git remote add origin https://github.com/zRitsu/disnake-LL-music-bot.git
+  git fetch origin
+  git checkout -b main -f --track origin/main
+else
+  git reset --hard
+  git pull --allow-unrelated-histories -X theirs
+fi
 
 if [ ! -d "venv" ]; then
   if [ -x "$(command -v py)" ]; then
@@ -17,9 +24,8 @@ if [ ! -d "venv" ]; then
     sleep 45
     exit 1
   fi
-
 fi
 
 source venv/Scripts/activate
 pip install -r requirements.txt --force-reinstall
-read -p "\n\nPressione ENTER para finalizar..."
+read -p "Pressione ENTER para finalizar..."
